Extracted test building blocks in shared project
