+++ test/NoEntityFrameworkTests/NoEntityFrameworkTests.csproj
@@ -11,12 +11,12 @@
 
   <ItemGroup>
     <ProjectReference Include="..\..\src\Examples\NoEntityFrameworkExample\NoEntityFrameworkExample.csproj" />
+    <ProjectReference Include="..\TestBuildingBlocks\TestBuildingBlocks.csproj" />
   </ItemGroup>
 
   <ItemGroup>
     <PackageReference Include="Microsoft.AspNetCore.Mvc.Testing" Version="$(AspNetCoreVersion)" />
     <PackageReference Include="Moq" Version="$(MoqVersion)" />
-    <PackageReference Include="xunit" Version="$(XUnitVersion)" />
     <PackageReference Include="xunit.runner.visualstudio" Version="$(XUnitVersion)" />
   </ItemGroup>
 </Project>
