+++ test/TestBuildingBlocks/FrozenSystemClock.cs
@@ -3,7 +3,7 @@ using Microsoft.AspNetCore.Authentication;
 
 namespace JsonApiDotNetCoreExampleTests
 {
-    internal sealed class FrozenSystemClock : ISystemClock
+    public sealed class FrozenSystemClock : ISystemClock
     {
         private static readonly DateTimeOffset _defaultTime =
             new DateTimeOffset(new DateTime(2000, 1, 1, 1, 1, 1), TimeSpan.FromHours(1));
