+++ test/UnitTests/Models/ResourceConstructionExpressionTests.cs
@@ -3,6 +3,7 @@ using System.ComponentModel.Design;
 using System.Linq.Expressions;
 using JsonApiDotNetCore.Resources;
 using JsonApiDotNetCoreExample.Data;
+using JsonApiDotNetCoreExampleTests;
 using Microsoft.AspNetCore.Authentication;
 using Microsoft.EntityFrameworkCore;
 using Xunit;
