+++ test/UnitTests/Models/ResourceConstructionTests.cs
@@ -5,6 +5,7 @@ using JsonApiDotNetCore.Middleware;
 using JsonApiDotNetCore.Resources;
 using JsonApiDotNetCore.Serialization;
 using JsonApiDotNetCoreExample.Data;
+using JsonApiDotNetCoreExampleTests;
 using Microsoft.AspNetCore.Http;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging.Abstractions;
